changed admin-only functions and done respective tests refactor
